disease/requirements.txt
flask
flask-cors
torch
torchvision
transformers
Pillow
gdown
gunicorn